+++ emcc.py
@@ -476,6 +476,19 @@ def do_emscripten(infile, memfile):
   return outfile
 
 
+def ensure_archive_index(archive_file):
+  # Fastcomp linking works without archive indexes
+  if not shared.Settings.WASM_BACKEND:
+    return
+  stdout = run_process([shared.LLVM_NM, '--print-armap', archive_file], stdout=PIPE).stdout
+  stdout = stdout.strip()
+  if stdout.startswith('Archive map\n') or stdout.startswith('Archive index\n'):
+    return
+  shared.warning('%s: archive is missing an index; Use emar when creating libraries to ensure an index is created', archive_file)
+  shared.warning('%s: adding index', archive_file)
+  run_process([shared.LLVM_RANLIB, archive_file])
+
+
 #
 # Main run() function
 #
@@ -990,7 +1003,9 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         # TODO(sbc): Pass this and other flags through when using lld
         # link_flags.append((i, arg))
         newargs[i] = ''
-
+      elif arg == '-':
+        input_files.append((i, arg))
+        newargs[i] = ''
     newargs = [a for a in newargs if a]
 
     if has_dash_c or has_dash_S:
@@ -1268,6 +1283,9 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
     if shared.Settings.DEMANGLE_SUPPORT:
       shared.Settings.EXPORTED_FUNCTIONS += ['___cxa_demangle']
       forced_stdlibs.append('libc++abi')
+
+    if shared.Settings.FULL_ES3:
+      shared.Settings.FULL_ES2 = 1
 
     if shared.Settings.EMBIND:
       forced_stdlibs.append('libembind')
@@ -1883,9 +1901,6 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         return any(flag.startswith(x) for x in ('-l', '-L', '-Wl,'))
 
       compile_args = [a for a in newargs if a and not is_link_flag(a)]
-      if '-fPIC' in compile_args and not shared.Settings.RELOCATABLE:
-        shared.warning('ignoring -fPIC flag when not building with SIDE_MODULE or MAIN_MODULE')
-        compile_args.remove('-fPIC')
 
       # Bitcode args generation code
       def get_clang_command(input_files):
@@ -1950,8 +1965,12 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           if file_ending.endswith(OBJECT_FILE_ENDINGS):
             logger.debug('using object file: ' + input_file)
             temp_files.append((i, input_file))
-          elif file_ending.endswith(DYNAMICLIB_ENDINGS) or shared.Building.is_ar(input_file):
-            logger.debug('using library file: ' + input_file)
+          elif file_ending.endswith(DYNAMICLIB_ENDINGS):
+            logger.debug('using shared library: ' + input_file)
+            temp_files.append((i, input_file))
+          elif shared.Building.is_ar(input_file):
+            logger.debug('using static library: ' + input_file)
+            ensure_archive_index(input_file)
             temp_files.append((i, input_file))
           elif file_ending.endswith(ASSEMBLY_ENDINGS):
             if not LEAVE_INPUTS_RAW:
@@ -2999,19 +3018,17 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
         options.binaryen_passes += ['--pass-arg=emscripten-sbrk-val@%d' % shared.Settings.DYNAMIC_BASE]
     if DEBUG:
       shared.safe_copy(wasm_binary_target, os.path.join(shared.get_emscripten_temp_dir(), os.path.basename(wasm_binary_target) + '.pre-byn'))
-    cmd = shared.Building.get_wasm_opt_command(wasm_binary_target,
-                                               wasm_binary_target,
-                                               options.binaryen_passes,
-                                               debug=intermediate_debug_info)
+    args = options.binaryen_passes
     if use_source_map(options):
-      cmd += ['--input-source-map=' + wasm_source_map_target]
-      cmd += ['--output-source-map=' + wasm_source_map_target]
-      cmd += ['--output-source-map-url=' + options.source_map_base + os.path.basename(wasm_binary_target) + '.map']
+      args += ['--input-source-map=' + wasm_source_map_target]
+      args += ['--output-source-map=' + wasm_source_map_target]
+      args += ['--output-source-map-url=' + options.source_map_base + os.path.basename(wasm_binary_target) + '.map']
       if DEBUG:
         shared.safe_copy(wasm_source_map_target, os.path.join(shared.get_emscripten_temp_dir(), os.path.basename(wasm_source_map_target) + '.pre-byn'))
-    logger.debug('wasm-opt on binaryen passes: %s', cmd)
-    shared.print_compiler_stage(cmd)
-    shared.check_call(cmd)
+    shared.Building.run_wasm_opt(wasm_binary_target,
+                                 wasm_binary_target,
+                                 args=args,
+                                 debug=intermediate_debug_info)
   if shared.Settings.BINARYEN_SCRIPTS:
     binaryen_scripts = os.path.join(shared.BINARYEN_ROOT, 'scripts')
     script_env = os.environ.copy()
@@ -3092,7 +3109,7 @@ def do_binaryen(target, asm_target, options, memfile, wasm_binary_target,
   # this will also remove debug info if we only kept it around in the intermediate invocations.
   # note that wasm2js handles the symbol map itself (as it manipulates and then
   # replaces the wasm with js)
-  if intermediate_debug_info and not shared.Settings.WASM2JS:
+  if options.emit_symbol_map and not shared.Settings.WASM2JS:
     shared.Building.handle_final_wasm_symbols(wasm_file=wasm_binary_target, symbols_file=symbols_file, debug_info=debug_info)
     save_intermediate_with_wasm('symbolmap', wasm_binary_target)